Update versioning plugin

+++ evolveum-jekyll-plugin/lib/evolveum/jekyll-versioning-plugin.rb
@@ -9,7 +9,11 @@ $stdout.reopen("/var/log/jekylversioning", "w")
 
 
 def installVersions(versions)
-  IO.popen("rm -rf /docs/midpoint/reference/*")
+  if Dir["/mp-#{versions[0]}"].empty?
+    IO.popen("rm -rf /docs/midpoint/reference/*")
+    IO.popen("cd /docs && grep -rl midpoint/reference . | xargs sed -i '/\/master/! s/midpoint\/reference/midpoint\/reference\/master/'")
+  end
+
   versions.each do |version|
     if Dir["/mp-#{version}"].empty?
       IO.popen("cd / && git clone -b #{version} https://github.com/janmederly/testversioning mp-#{version} && rm /mp-#{version}/docs/LICENSE && ln -s /mp-#{version}/docs/ /docs/midpoint/reference/#{version}") #maybe
